requirements.txt
asgiref==3.7.2
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
fonttools==4.43.1
idna==3.4
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.0
matplotlib==3.8.0
msgpack==1.0.7
numba==0.58.0
numpy==1.25.2
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
psycopg2-binary==2.9.9
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.0.6
whitenoise==6.6.0
zipp==3.17.0
